26/3_mtu_capnhatdiachi_nguoidung

+++ SneakFit.Application/Catalog/SanPhamChiTiet/SanPhamChiTetService.cs
@@ -106,10 +106,10 @@ namespace SneakFit.Application.Catalog.SanPhamChiTietChiTiet
                 query = query.Where(x => x.sp.DanhMucId == request.DanhMucId);
 
             if (request.GiaThapNhat.HasValue)
-                query = query.Where(x => x.spct.Gia >= (float)request.GiaThapNhat.Value);
+                query = query.Where(x => x.spct.Gia >= request.GiaThapNhat.Value);
 
             if (request.GiaCaoNhat.HasValue)
-                query = query.Where(x => x.spct.Gia <= (float)request.GiaCaoNhat.Value);
+                query = query.Where(x => x.spct.Gia <= request.GiaCaoNhat.Value);
 
             if (request.LocTrangthai == true)
                 query = query.Where(x => x.spct.TrangThai == request.TrangThai);
